+++ CHANGELOG.md
@@ -11,5 +11,7 @@
 - Añadido método `nuevoUsuario` a ``ServicioIncidencia`` para registrar un nuevo `Usuario` en el sistema (Alberto Jiménez Expósito)
 - Añadido método ``login`` a `ServicioIncidencia` para hacer login de un usuario ya registrado en el sistema (Alberto Jiménez Expósito)
 - Inicialización de ``tiposIncidencia`` en `ServicioIncidencia` para añadir los tipos de incidencia por defecto en el sistema (Alberto Jiménez Expósito)
+- Añadido método ``obtenerListaIncidenciasUsuario`` a `ServicioIncidencia` para buscar incidencias por email (Alberto Jiménez Expósito)
 - Añadida clase ``EstadoIncidencia`` que sirve para indicar los estados posibles de una `Incidencia` (Carlos Mayor Navarro)
 - Modificaciones en la declaración del atributo``estado`` dentro de `Incidencia` y métodos getter y setter, además del resto métodos relacionados con la misma en ``ServicioIncidencia`` (Carlos Mayor Navarro)
+- Modificación del método ``buscarIncidenciasTipoEstado`` en `ServicioIncidencias` para simplificar su lógica (Alberto Jiménez Expósito)